+++ src/Phan/Language/Element/Clazz.php
@@ -14,6 +14,7 @@ use Phan\Exception\IssueException;
 use Phan\Issue;
 use Phan\IssueFixSuggester;
 use Phan\Language\Context;
+use Phan\Language\Element\Comment\Property as CommentProperty;
 use Phan\Language\FQSEN;
 use Phan\Language\FQSEN\FullyQualifiedClassConstantName;
 use Phan\Language\FQSEN\FullyQualifiedClassName;
@@ -338,7 +339,6 @@ class Clazz extends AddressableElement
         }
         $class_fqsen = $clazz->getFQSEN();
         $class_fqsen_string = strtolower($class_fqsen->__toString());
-        ;
         if ($class_fqsen_string === '\closure') {
             $parameter_list = [
                 new Parameter($clazz->getContext(), 'callable', UnionType::fromFullyQualifiedString('callable'), 0)
@@ -446,7 +446,7 @@ class Clazz extends AddressableElement
             throw new \Exception("Class $this has no parent");
         }
 
-        return $parent_type_option->get()->asFQSEN();
+        return FullyQualifiedClassName::fromType($parent_type_option->get());
     }
 
     /**
@@ -615,7 +615,6 @@ class Clazz extends AddressableElement
     /**
      * @return array<int,FullyQualifiedClassName>
      * Get the list of interfaces implemented by this class
-     * @suppress PhanPartialTypeMismatchReturn
      */
     public function getInterfaceFQSENList() : array
     {
@@ -760,7 +759,7 @@ class Clazz extends AddressableElement
     }
 
     /**
-     * @param array<string,\Phan\Language\Element\Comment\Parameter> $magic_property_map mapping from property name to property
+     * @param array<string,CommentProperty> $magic_property_map mapping from property name to property
      * @param CodeBase $code_base
      * @return bool whether or not we defined it.
      */
@@ -785,7 +784,7 @@ class Clazz extends AddressableElement
                 $property_name
             );
             $property = new Property(
-                $context,
+                clone($context)->withLineNumberStart($comment_parameter->getLine()),
                 $property_name,
                 $comment_parameter->getUnionType(),
                 $flags,
@@ -834,7 +833,7 @@ class Clazz extends AddressableElement
                 return $parameter->asRealParameter($context);
             }, $comment_method->getParameterList());
             $method = new Method(
-                $context,
+                clone($context)->withLineNumberStart($comment_method->getLine()),
                 $method_name,
                 $comment_method->getUnionType(),
                 $flags,
@@ -1705,7 +1704,6 @@ class Clazz extends AddressableElement
     /**
      * @return array<int,FullyQualifiedClassName>
      * A list of FQSEN's for included traits
-     * @suppress PhanPartialTypeMismatchReturn TODO: investigate
      */
     public function getTraitFQSENList() : array
     {
@@ -2500,7 +2498,7 @@ class Clazz extends AddressableElement
 
         $parent_type = $this->parent_type;
         if ($parent_type) {
-            $extend_types[] = $parent_type->asFQSEN();
+            $extend_types[] = FullyQualifiedClassName::fromType($parent_type);
             $parent_class = $this->getParentClass($code_base);
             $parent_implements_types = $parent_class->interface_fqsen_list;
         }