+++ src/Phan/Language/FQSEN/FullyQualifiedGlobalStructuralElement.php
@@ -6,10 +6,10 @@ use Phan\Language\Context;
 use Phan\Language\Type;
 
 use AssertionError;
+use InvalidArgumentException;
 
 /**
  * A Fully-Qualified Global Structural Element
- * @phan-file-suppress PhanPluginNoAssert
  */
 abstract class FullyQualifiedGlobalStructuralElement extends AbstractFQSEN
 {
@@ -35,6 +35,12 @@ abstract class FullyQualifiedGlobalStructuralElement extends AbstractFQSEN
      *
      * @throws EmptyFQSENException
      * if the name component of this FullyQualifiedGlobalStructuralElement is empty
+     *
+     * @throws EmptyFQSENException
+     * if the namespace of this FullyQualifiedGlobalStructuralElement is empty
+     *
+     * @throws InvalidArgumentException
+     * if the namespace begins with an invalid character
      */
     protected function __construct(
         string $namespace,
@@ -45,12 +51,13 @@ abstract class FullyQualifiedGlobalStructuralElement extends AbstractFQSEN
             throw new EmptyFQSENException("The name of an FQSEN cannot be empty", rtrim($namespace, '\\') . '\\');
         }
 
-        \assert($namespace !== '', "The namespace cannot be empty");
+        if ($namespace === '') {
+            throw new EmptyFQSENException("The namespace cannot be empty", $name);
+        }
 
-        \assert(
-            $namespace[0] === '\\',
-            "The first character of a namespace must be \\"
-        );
+        if ($namespace[0] !== '\\') {
+            throw new InvalidArgumentException("The first character of a namespace must be \\");
+        }
 
         parent::__construct($name);
         $this->namespace = $namespace;
@@ -197,7 +204,9 @@ abstract class FullyQualifiedGlobalStructuralElement extends AbstractFQSEN
         // Split the parts into the namespace(0 or more components) and the last name.
         $name = \array_pop($parts);
 
-        \assert(!empty($name), "The name cannot be empty");
+        if (empty($name)) {
+            throw new AssertionError("The name cannot be empty");
+        }
 
         // Check for a name map
         if ($context->hasNamespaceMapFor($namespace_map_type, $fqsen_string)) {
@@ -275,10 +284,9 @@ abstract class FullyQualifiedGlobalStructuralElement extends AbstractFQSEN
             return $this;
         }
 
-        \assert(
-            $alternate_id < 1000,
-            "Your alternate IDs have run away"
-        );
+        if ($alternate_id >= 1000) {
+            throw new AssertionError("Your alternate IDs have run away");
+        }
 
         return static::make(
             $this->getNamespace(),