+++ src/Phan/Language/Scope.php
@@ -10,9 +10,8 @@ use Phan\Language\FQSEN\FullyQualifiedMethodName;
 use Phan\Language\FQSEN\FullyQualifiedPropertyName;
 use Phan\Language\Type\TemplateType;
 
-/**
- * @phan-file-suppress PhanPluginNoAssert
- */
+use AssertionError;
+
 abstract class Scope
 {
     /**
@@ -62,6 +61,7 @@ abstract class Scope
     /**
      * @return Scope
      * Get the parent scope of this scope
+     * @suppress PhanPossiblyNullTypeReturn callers should call hasParentScope
      */
     public function getParentScope() : Scope
     {
@@ -80,6 +80,7 @@ abstract class Scope
 
     /**
      * @return FQSEN
+     * @suppress PhanPossiblyNullTypeReturn callers should call hasFQSEN
      */
     public function getFQSEN()
     {
@@ -102,10 +103,9 @@ abstract class Scope
      */
     public function getClassFQSEN() : FullyQualifiedClassName
     {
-        \assert(
-            $this->hasParentScope(),
-            "Cannot get class FQSEN on scope"
-        );
+        if (!$this->hasParentScope()) {
+            throw new AssertionError("Cannot get class FQSEN on scope");
+        }
 
         return $this->getParentScope()->getClassFQSEN();
     }
@@ -126,10 +126,9 @@ abstract class Scope
      */
     public function getPropertyFQSEN() : FullyQualifiedPropertyName
     {
-        \assert(
-            $this->hasParentScope(),
-            "Cannot get class FQSEN on scope"
-        );
+        if (!$this->hasParentScope()) {
+            throw new AssertionError("Cannot get class FQSEN on scope");
+        }
 
         return $this->getParentScope()->getPropertyFQSEN();
     }
@@ -150,10 +149,9 @@ abstract class Scope
      */
     public function getFunctionLikeFQSEN()
     {
-        \assert(
-            $this->hasParentScope(),
-            "Cannot get method/function/closure FQSEN on scope"
-        );
+        if (!$this->hasParentScope()) {
+            throw new AssertionError("Cannot get method/function/closure FQSEN on scope");
+        }
 
         return $this->getParentScope()->getFunctionLikeFQSEN();
     }
@@ -250,10 +248,9 @@ abstract class Scope
      */
     public function addGlobalVariable(Variable $variable)
     {
-        \assert(
-            $this->hasParentScope(),
-            "No global scope available. This should not happen."
-        );
+        if (!$this->hasParentScope()) {
+            throw new AssertionError("No global scope available. This should not happen.");
+        }
 
         $this->getParentScope()->addGlobalVariable($variable);
     }
@@ -264,14 +261,11 @@ abstract class Scope
      */
     public function hasGlobalVariableWithName(string $name) : bool
     {
-        \assert(
-            $this->hasParentScope(),
-            "No global scope available. This should not happen."
-        );
+        if (!$this->hasParentScope()) {
+            throw new AssertionError("No global scope available. This should not happen.");
+        }
 
-        return $this->getParentScope()->hasGlobalVariableWithName(
-            $name
-        );
+        return $this->getParentScope()->hasGlobalVariableWithName($name);
     }
 
     /**
@@ -280,10 +274,9 @@ abstract class Scope
      */
     public function getGlobalVariableByName(string $name) : Variable
     {
-        \assert(
-            $this->hasParentScope(),
-            "No global scope available. This should not happen."
-        );
+        if (!$this->hasParentScope()) {
+            throw new AssertionError("No global scope available. This should not happen.");
+        }
 
         return $this->getParentScope()->getGlobalVariableByName($name);
     }
@@ -356,10 +349,9 @@ abstract class Scope
         string $template_type_identifier
     ) : TemplateType {
 
-        \assert(
-            $this->hasTemplateType($template_type_identifier),
-            "Cannot get template type with identifier $template_type_identifier"
-        );
+        if (!$this->hasTemplateType($template_type_identifier)) {
+            throw new AssertionError("Cannot get template type with identifier $template_type_identifier");
+        }
 
         return $this->template_type_map[$template_type_identifier]
             ?? $this->getParentScope()->getTemplateType(